fix sudo issue
